Add donated scratch regs to depency list only when they are used
Adds msrDonateUsed flag to mark that a donated scratch register has been
used.
addScratchRegistersToDependencyList will not add a donated regsiter to
the dependency list unless the msrDonateUsed flag is set.

This fixes an issue where a register is donated but not used and is
still added to be dependency list, causing an extra register to be
assigned at labels marked with endInternalControlFlow

Signed-off-by: Matthew Hall <matthew.hall3@outlook.com>